requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
boto3==1.27.1
botocore==1.30.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
cryptography==40.0.2
defusedxml==0.7.1
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.2
flatbuffers==23.5.26
fonttools==4.40.0
fpdf==1.7.2
frozenlist==1.3.3
gunicorn==21.2.0
h11==0.14.0
humanfriendly==10.0
idna==3.4
imageio==2.31.1
jmespath==1.0.1
lazy_loader==0.3
llvmlite==0.40.1
Mako==1.2.4
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.24.3
onnxruntime==1.15.1
openai==0.27.8
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.0.1
passlib==1.7.4
pdfkit==1.0.0
Pillow==10.0.0
platformdirs==3.10.0
pooch==1.7.0
protobuf==4.24.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
PyMatting==1.1.8
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
rembg==2.0.50
removebg==0.4
reportlab==4.0.4
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
scikit-image==0.21.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
sympy==1.12
tifffile==2023.8.12
tqdm==4.65.0
typing_extensions==4.6.2
tzdata==2023.3
ua-parser==0.18.0
urllib3==1.26.16
user-agent==0.1.10
user-agents==2.2.0
uvicorn==0.22.0
yarl==1.9.2
